add date added support to sqlite entry mappings manager
